Fixed url to prevent updates causing broken links
A more specific url has been provided to resolve an issue where a new
release of a particular software would cause a 404 error when the script
attempted to download an older version of the software. The url that the
script attempts to use to download is now specific to the software
version specified in the configuration file.