Added hostname validation features
Added validation of hostname through the addition of a check function,
validate_hostname() and an if statement checking for the return value of
aforementioned function. If the hostname is valid, then the user
supplied value is assigned to the variable. Otherwise, script will
abort. This is necessary to filter out bad values that would otherwise
cause errors that would be mostly invisible to the user without further
investigation.